+++ BE/girin-grim/src/main/java/com/starshop/giringrim/funding/service/FundingService.java
@@ -28,4 +28,5 @@ public interface FundingService {
      */
     FundingRespDtos.HomeDto home(Integer page, Long universityId, String fundingType, String keyword, String method, UserDetailsImpl userDetails);
 
+
 }
