+++ BE/girin-grim/src/test/java/com/starshop/giringrim/funding/controller/FundingControllerTest.java
@@ -1,7 +1,6 @@
 package com.starshop.giringrim.funding.controller;
 
-import com.starshop.giringrim.TestConfig;
-import com.starshop.giringrim.funding.dto.FundingRespDtos;
+import com.starshop.giringrim.config.TestConfig;
 import net.minidev.json.JSONObject;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -10,8 +9,6 @@ import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.jdbc.Sql;
-import org.springframework.test.context.jdbc.SqlConfig;
 
 import static org.assertj.core.api.Assertions.assertThat;
